(feat): Transformed project to OOP
